Move map info box components and consolidate styles
